Moved LDP, RDF, and RDFS vocabularies to *vocabulary* package within the API
