+++ framework/application/api/src/main/java/org/ldp4j/application/ext/annotations/MembershipRelation.java
@@ -28,7 +28,7 @@ package org.ldp4j.application.ext.annotations;
 
 import java.net.URI;
 
-import org.ldp4j.application.domain.LDP;
+import org.ldp4j.application.vocabulary.LDP;
 import org.ldp4j.application.vocabulary.Term;
 
 public enum MembershipRelation {
