+++ framework/application/core/src/main/java/org/ldp4j/application/DefaultPublicIndirectContainer.java
@@ -32,13 +32,13 @@ import org.ldp4j.application.data.DataSet;
 import org.ldp4j.application.data.Individual;
 import org.ldp4j.application.data.validation.ValidationConstraintFactory;
 import org.ldp4j.application.data.validation.Validator.ValidatorBuilder;
-import org.ldp4j.application.domain.LDP;
 import org.ldp4j.application.endpoint.Endpoint;
 import org.ldp4j.application.engine.context.ContentPreferences;
 import org.ldp4j.application.engine.context.PublicIndirectContainer;
 import org.ldp4j.application.engine.context.PublicResource;
 import org.ldp4j.application.engine.context.PublicResourceVisitor;
 import org.ldp4j.application.template.IndirectContainerTemplate;
+import org.ldp4j.application.vocabulary.LDP;
 import org.ldp4j.application.vocabulary.Term;
 
 final class DefaultPublicIndirectContainer extends DefaultPublicMembershipAwareContainer<IndirectContainerTemplate> implements PublicIndirectContainer {
