+++ framework/server/command/src/main/java/org/ldp4j/server/resources/impl/IndirectContainerImpl.java
@@ -29,7 +29,7 @@ package org.ldp4j.server.resources.impl;
 import java.net.URI;
 
 import org.ldp4j.application.data.Individual;
-import org.ldp4j.application.domain.LDP;
+import org.ldp4j.application.vocabulary.LDP;
 import org.ldp4j.server.resources.IndirectContainer;
 import org.ldp4j.server.resources.ResourceType;
