+++ framework/server/command/src/main/java/org/ldp4j/server/resources/impl/ResourceImpl.java
@@ -41,9 +41,9 @@ import org.ldp4j.application.data.ManagedIndividual;
 import org.ldp4j.application.data.ManagedIndividualId;
 import org.ldp4j.application.data.Name;
 import org.ldp4j.application.data.Value;
-import org.ldp4j.application.domain.LDP;
-import org.ldp4j.application.domain.RDF;
 import org.ldp4j.application.resource.ResourceId;
+import org.ldp4j.application.vocabulary.LDP;
+import org.ldp4j.application.vocabulary.RDF;
 import org.ldp4j.application.vocabulary.Term;
 import org.ldp4j.server.Entity;
 import org.ldp4j.server.resources.Resource;
