+++ framework/server/command/src/test/java/org/ldp4j/server/impl/TurtleMediaTypeProviderTest.java
@@ -50,9 +50,9 @@ import org.ldp4j.application.data.ManagedIndividualId;
 import org.ldp4j.application.data.Name;
 import org.ldp4j.application.data.NameVisitor;
 import org.ldp4j.application.data.NamingScheme;
-import org.ldp4j.application.domain.LDP;
-import org.ldp4j.application.domain.RDF;
 import org.ldp4j.application.resource.ResourceId;
+import org.ldp4j.application.vocabulary.LDP;
+import org.ldp4j.application.vocabulary.RDF;
 import org.ldp4j.application.vocabulary.Term;
 import org.ldp4j.rdf.URIRef;
 import org.ldp4j.server.ResourceIndex;
