+++ framework/server/core/src/main/java/org/ldp4j/server/controller/EndpointControllerUtils.java
@@ -34,7 +34,6 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response.ResponseBuilder;
 import javax.ws.rs.core.Variant;
 
-import org.ldp4j.application.domain.LDP;
 import org.ldp4j.application.engine.context.Capabilities;
 import org.ldp4j.application.engine.context.PublicBasicContainer;
 import org.ldp4j.application.engine.context.PublicDirectContainer;
@@ -42,6 +41,7 @@ import org.ldp4j.application.engine.context.PublicIndirectContainer;
 import org.ldp4j.application.engine.context.PublicRDFSource;
 import org.ldp4j.application.engine.context.PublicResource;
 import org.ldp4j.application.engine.context.PublicResourceVisitor;
+import org.ldp4j.application.vocabulary.LDP;
 import org.ldp4j.application.vocabulary.Term;
 import org.ldp4j.server.spi.RuntimeInstance;
