+++ framework/server/core/src/main/java/org/ldp4j/server/controller/ExistingEndpointController.java
@@ -39,9 +39,6 @@ import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.Variant;
 
 import org.ldp4j.application.data.DataSet;
-import org.ldp4j.application.domain.LDP;
-import org.ldp4j.application.domain.RDF;
-import org.ldp4j.application.domain.RDFS;
 import org.ldp4j.application.engine.context.ApplicationContextException;
 import org.ldp4j.application.engine.context.ApplicationExecutionException;
 import org.ldp4j.application.engine.context.ContentPreferences;
@@ -58,6 +55,9 @@ import org.ldp4j.application.ext.InconsistentContentException;
 import org.ldp4j.application.ext.InvalidContentException;
 import org.ldp4j.application.ext.UnknownResourceException;
 import org.ldp4j.application.ext.UnsupportedContentException;
+import org.ldp4j.application.vocabulary.LDP;
+import org.ldp4j.application.vocabulary.RDF;
+import org.ldp4j.application.vocabulary.RDFS;
 import org.ldp4j.rdf.Namespaces;
 import org.ldp4j.server.utils.VariantUtils;
 import org.slf4j.Logger;
